Reset 'PROYECTO-CURSO-Cruce_Controlado_v0.ino'

+++ 13b2-delayed_fading_light_off_timed_millis/13b2-delayed_fading_light_off_timed_millis.ino
@@ -12,7 +12,7 @@ int OffTime = 3000;
 int FadingTime = 2000;
 
 int timerON, timerFADING, timerOFF;
-byte brillo;  
+int brillo;  
 
 void setup() {
   pinMode(LED_PIN, OUTPUT);
